MM-43 continuing to organize components and fix errors
