--- jiraDemo
@@ -1,2 +1,0 @@
-change for first commit! hello!
-jira test.
