+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import store from './redux/store';
 import rootSaga from './redux/sagas/_root.saga';
-import App from './components/App/App';
+import App from '../../many-murals/src/components/Project/App/App';
 
 ReactDOM.render(
   <Provider store={store}>
